pyproject.toml
[project]
name = "masterclass-agentes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "langchain-tavily>=0.2.11",
    "langchain[openai]>=0.3.27",
    "langgraph>=0.5.4",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "pycryptodome>=3.23.0",
    "python-dotenv>=1.1.1",
]
